refactor(api): modularize services and hooks into domain-specific files
- Created domain-specific service files in src/api/services/ (auth, assets, users, etc.).
- Created domain-specific hook files in src/hooks/api/.
- Extracted shared useAsyncOperation helper for consistent API interaction.
- Updated src/api/services.js and src/hooks/useAssetTrack.js to act as proxies for the new modules.
- Updated Sidebar and TopNav components to use modernized modular hooks.
- Enhanced API client with a fallback for environment variables in testing environments.
