+++ frontend/src/api/services/allocations.js
@@ -1,0 +1,10 @@
+import { get, post } from '../client';
+
+export const allocationService = {
+  history: (assetId, params) => get(`/assets/${assetId}/allocations`, { params }),
+  allocate: (assetId, body) => post(`/assets/${assetId}/allocations`, body),
+  getById: (assetId, allocationId) => get(`/assets/${assetId}/allocations/${allocationId}`),
+  deallocate: (assetId, allocationId, body) => post(`/assets/${assetId}/allocations/${allocationId}/deallocate`, body),
+};
+
+export default allocationService;
