+++ frontend/src/api/services/index.js
@@ -1,0 +1,9 @@
+export { authService } from './auth';
+export { userService } from './users';
+export { assetService } from './assets';
+export { allocationService } from './allocations';
+export { conditionReportService } from './conditions';
+export { dashboardService } from './dashboard';
+export { notificationService } from './notifications';
+export { searchService } from './search';
+export { reportService } from './reports';
