+++ frontend/src/hooks/api/useAllocations.js
@@ -1,0 +1,12 @@
+import { allocationService } from '../../api/services/allocations';
+import useAsyncOperation from './useAsyncOperation';
+
+export const useAllocateAsset = () => {
+  const { execute, loading, error, clearError } = useAsyncOperation(allocationService.allocate);
+  return { allocate: execute, loading, error, clearError };
+};
+
+export const useDeallocateAsset = () => {
+  const { execute, loading, error, clearError } = useAsyncOperation(allocationService.deallocate);
+  return { deallocate: execute, loading, error, clearError };
+};
