+++ frontend/src/hooks/api/useAuth.js
@@ -1,0 +1,45 @@
+import { useCallback } from 'react';
+import { authService } from '../../api/services/auth';
+import { storeToken } from '../../api/client';
+import useAuthStore from '../../store/useAuthStore';
+import useAsyncOperation from './useAsyncOperation';
+
+export const useLogin = () => {
+  const { execute, loading, error, clearError } = useAsyncOperation(authService.login);
+  const loginStore = useAuthStore((state) => state.login);
+
+  const login = useCallback(
+    async (credentials) => {
+      const authResponse = await execute(credentials);
+      if (authResponse) {
+        const { accessToken, user } = authResponse;
+        storeToken(accessToken);
+        loginStore(user, accessToken);
+      }
+      return authResponse;
+    },
+    [execute, loginStore],
+  );
+
+  return { login, loading, error, clearError };
+};
+
+export const useSignup = () => {
+  const { execute, loading, error, clearError } = useAsyncOperation(authService.signUp);
+
+  const signup = useCallback(
+    async (credentials) => execute(credentials),
+    [execute],
+  );
+
+  return { signup, loading, error, clearError };
+};
+
+export const useAuth = () => {
+  const user = useAuthStore((state) => state.user);
+  const token = useAuthStore((state) => state.token);
+  const isAuthenticated = useAuthStore((state) => state.isAuthenticated);
+  const logout = useAuthStore((state) => state.logout);
+
+  return { user, token, isAuthenticated, logout };
+};
